Refactor search index construction to search index type
