+++ backend/src/search/SearchIndex.ts
@@ -6,12 +6,87 @@ import {
   SearchResult,
 } from '../shared/search';
 import { IndexEntry } from './IndexEntry';
+import { SearchIndexEntryRecorder } from './SearchIndexEntryRecorder';
+import { AppService } from '../app.service';
+import { SchemaService } from '../schema/schema.service';
 
 /**
  * Represents a search index that can be used to perform search operations and auto suggestions.
  */
 export class SearchIndex {
-  public constructor(private readonly index: AsyncLazy<MiniSearch>) {}
+  private constructor(private readonly index: AsyncLazy<MiniSearch>) {}
+
+  public static create(
+    queryService: AppService,
+    schemaService: SchemaService
+  ): SearchIndex {
+    return new SearchIndex(
+      new AsyncLazy<MiniSearch>(() =>
+        SearchIndex.createCore(queryService, schemaService)
+      )
+    );
+  }
+
+  private static async createCore(
+    queryService: AppService,
+    schemaService: SchemaService
+  ): Promise<MiniSearch> {
+    const entryRecorder = new SearchIndexEntryRecorder();
+
+    // TODO: For large databases it would be beneficial, if we read the entries in smaller batches,
+    //       so we don't have to copy the entire dataset into memory (again).
+    //       The problem here, is that we need the property names before adding the entries to the
+    //       index. This can be done by executing an up-front query for the schema of the data,
+    //       as is done in the SchemaService.
+
+    // Get all nodes from the database
+    const nodes = await queryService.getNodes();
+
+    for (const node of nodes) {
+      entryRecorder.recordNode(node);
+    }
+
+    // Get all edges from the database
+    const edges = await queryService.getEdges();
+
+    for (const edge of edges) {
+      entryRecorder.recordEdge(edge);
+    }
+
+    // Get all node types from the database
+    const nodeTypes = await schemaService.getNodeTypes();
+
+    for (const nodeType of nodeTypes) {
+      entryRecorder.recordNodeType(nodeType);
+    }
+
+    // Get all edge types from the database
+    const edgeTypes = await schemaService.getEdgeTypes();
+
+    for (const edgeType of edgeTypes) {
+      entryRecorder.recordEdgeType(edgeType);
+    }
+
+    // Build the index
+    const fields = ['indexValue'];
+    const storeFields = [
+      'entityType',
+      'id',
+      'types',
+      'from',
+      'to',
+      'indexKey',
+      'indexValue',
+    ];
+
+    const index = new MiniSearch({
+      idField: 'id',
+      fields,
+      storeFields,
+    });
+    index.addAll(entryRecorder.recordedEntries);
+    return index;
+  }
 
   /**
    * Searches through all entities and entity types and returns those, that
